requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
duckduckgo-search>=6.1.0
pyyaml>=6.0.1
rich>=13.7.0
typer>=0.12.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
pydantic>=2.0.0
google-generativeai>=0.8.0
